+++ pkg/compactor/compactor.go
@@ -273,6 +273,10 @@ type MultitenantCompactor struct {
 	// TSDB syncer metrics
 	syncerMetrics *aggregatedSyncerMetrics
 
+	// Block upload metrics
+	blockUploadBlocks      *prometheus.GaugeVec
+	blockUploadBytes       *prometheus.GaugeVec
+	blockUploadFiles       *prometheus.GaugeVec
 	blockUploadValidations atomic.Int64
 }
 
@@ -367,6 +371,18 @@ func newMultitenantCompactor(
 			Help:        blocksMarkedForDeletionHelp,
 			ConstLabels: prometheus.Labels{"reason": "compaction"},
 		}),
+		blockUploadBlocks: promauto.With(registerer).NewGaugeVec(prometheus.GaugeOpts{
+			Name: "cortex_block_upload_api_blocks_total",
+			Help: "Total number of blocks successfully uploaded and validated using the block upload API.",
+		}, []string{"user"}),
+		blockUploadBytes: promauto.With(registerer).NewGaugeVec(prometheus.GaugeOpts{
+			Name: "cortex_block_upload_api_bytes_total",
+			Help: "Total number of bytes from successfully uploaded and validated blocks using block upload API.",
+		}, []string{"user"}),
+		blockUploadFiles: promauto.With(registerer).NewGaugeVec(prometheus.GaugeOpts{
+			Name: "cortex_block_upload_api_files_total",
+			Help: "Total number of files from successfully uploaded and validated blocks using block upload API.",
+		}, []string{"user"}),
 	}
 
 	promauto.With(registerer).NewGaugeFunc(prometheus.GaugeOpts{